fix #4 Entropy float config value issue

+++ config/webconfig.go
@@ -47,7 +47,7 @@ type PassPolicy struct {
 	Min              int
 	Max              int
 	AllowReadSSHA256 bool
-	Entropy          float64
+	Entropy          int64
 }
 
 type CfgUsers struct {
